Add several bits of debugging information.
Disable following a dlopen.  Rationale from email message:

It has been a while, and I forgot how the threaded data was stored.  I actually set things up so that PAPI keeps track of the buffers for each thread in TLS, a nice little PAPI feature.  The vector stuff is for multiple events.  No need for a dynamic vector.  And, yes, when libmonitor detects threads it invokes a "you've got threads" callback.

And so this leads into the problem: a dlopen was handed over to hpcrun's primitive dlopen code (a hack submitted as a patch).  This changed load map information for one thread, but not for the other.  So, when the 'other' thread went to write out its data, it attempted to write out the 'extra' load module buffers and found invalid pointers.

As a quick tourniquet, I've just disabled the dlopen code.  So hpcrun just ignores the newly loaded DSOs and hopes that things haven't been remapped (which I think should be good enough for these codes).  In the fullness of time, hpcrun clearly needs the notion of epochs.  And it would be nice to base csprof/hpcrun on libmonitor.