bloop: When CFG is built, switch statements (indirect jumps) are not handled correctly.  In particular, the jump is ignored and no edge is added from the block to its targets.
Improve handling of incomplete CFGs by marking nodes while processing
the loop nest tree and then post processing unmarked nodes.  This
allows statement level attribution, even if the statements are not
properly nested in loops.

Result: This had a surprising effect of increasing coverage of structure
information.  Small jump tables are used in several places.

Oddity: A set of VMAs A mapping to line x may be located within the
structure, but another set of VMA's B also mapping to line x may not.
In the case, if a VMA in set B is not a call site, it will be moved
inward by coalesce-duplicate-statements.  If, however, the VMA is a
call site, it will remain rooted at the procedure... We could add a
flag to indicate that we are uncertain of the location of VMAs in set
B and then not assign them a distinct sort id so that these VMA are
moved inward...