- Write lush CCT to disk.
- Move CCT writing code from from hpcfile_cstree.c to cct.c. With
LUSH, the cct writing is becoming more complicated.  Because, 1) to
support LUSH the callback interface is no longer sufficient; 2) we
want writing to be stremlined; and 3) there is no need for a CCT
feedback loop (as with experiment databases), I no longer think that
we benefit from an abstract tree writer.  I do think an abstract
reader is useful and it can be maintain its simple interface given a
reasonable format.