Create --enable-hpcrun-static option to enable/disable hpcrun's static profiling code.
The motivation for this option is that it can be very tricky to
build the static profiling code when cross-compiling for a
multilib system.

The problem is related to the fact that the final libtool target is
a .o (rather than a .la) linked together from other .o's.  Libtool
correctly creates the individual .o's.  But when the time comes to
link the individual .o's together, problems occur.  The link begins
promisingly using --mode=link and the cross compiler (e.g. scgcc32).
However, libtool then uses the linker (ld) directly.  Unfortunately,
the linker doesn't know about cross-compiling multilib magic and
uses the default ABI rather than the correct one (e.g., 64 instead
of n32).  Even worse, trying to solve the problem using by setting
LDFLAGS (e.g. LDFLAGS="-Wl,-m...") appears futile because the flag
is not passed to ld when it is called directly.