Add unwind step validation for x86 unwinds:
   1) Always for trolling
   2) Optionally for recipe steps

Currently, anything with tail calls (including plt calls) is categorized
as "Probable".

The recipe-step validation is activated by

  -dd UNW_VALID

A summary of confirmed, probable and wrong steps is printed at the end, along
with the other summary information. Wrong unwinds are also listed in the log.

Additionally,

  -dd VALID_RECORD_ALL

may be added. This -dd flag causes *all* unwind steps' validation status to be
logged. USE IT SPARINGLY. It generates extremely large logs.

A dummy routine for validation summary is included for NON-x86 unwinds.