- Develop a work around for the following problem in sample_source::papi::papi_event_handler:
This code assumes that the metric id obtained from csprof_new_metric() is identical to the index of an event in a PAPI event set.  In truth, there is no logical connection between the two, even though they are used interchangeably.  Moreover, there is no check to ensure that the assumption holds.  With logical unwinding this house of cards falls down.

- Harden and clean the code that creates LUSH idleness metric.