Add new sample source SYNC, to accomodate user-defined synchronous samples in dynamic case.
Also add user-defined custom initialization to dynamic
case.

Add example of introspective custom metrics to T1 directory
(can be used as a prototype for other apps that want to
do custom synchronous-sampling metrics). The example code
is a skeleton for run-time escape analysis.
A sample run script (chrun) is included as well.