Set the symtabAPI platform based on the object file wordsize (whether CC produces 32- or 64-bit objects) instead of the host wordsize.  This allows building a 32-bit version for a 64-bit host.  This only applies to x86-64 and powerpc.
Libunwind does not build cleanly in 32-bit mode on x86-64 due to some
cross-compile issues, so disable that config for now.  This only
resets the default, it can be reenabled via --with-libunwind.

Reset libunwind's install libdir to lib instead of lib64.  This was
causing problems on powerpc in 64-bit mode.  Libunwind fails to obey
--libdir in this case, so the simplest solution is to patch configure.

Some packages still have problems building in 32-bit mode on 64-bit
hosts due to libiberty and other issues.