Rework how we open files in hpcrun.  Open and rename all of the files in files.c and choose all of the file names there.  Everything else just uses a file descriptor.
The hard part is to guarantee unique file names across all processes
and nodes.  Normally, (hostid, pid, gen) is sufficient.  However, on
some systems, gethostid() or getpid() do not return unique values.
Plus, a process that exec()s the same binary will always have the same
hostid and pid.

The solution is to open() the file with O_EXCL and choose a different
name if open() fails.  The .log file is process wide and we use this
file as a lock.  The process that opens the .log file wins the race
and owns the (hostid, pid, gen) id for its file names.

If open fails several times, then give up and start picking random
hostids.  If this happens, it's likely that gethostid() is broken on
this system and this is a more robust way of handling this case.

It would also make sense to replace (hostid, pid, gen) with a single
random number of some length.