Add preliminary support for the IO sample source.  This source adds two metrics: the number of IO bytes read and written and covers both stream and unbuffered functions.
For slow or blocking functions, we record samples before and after the
function.  If a process blocks in kernel, then it won't receive async
interrupts and this may under report the time in the trace.  Using two
samples assures that we see the full span of the function in the trace
viewer.

This is still early support and so far, we only cover fread() and
fwrite().  The main problem is that overriding functions that we use
in our code (read, write) would lead to deadlock.  Before doing that,
I first need to generalize the async blocks to cover both sync and
async entry and exit points.