In the start/stop interface, make hpctoolkit_sampling_start() and stop() also start and stop the sample sources in the current thread.
This allows the application to put the stop and start functions
arround a region of code that is sensitive to taking signals.  Some
Gemini interconnect (GNI) functions called from inside gasnet_init()
or MPI_Init() on the Cray XE (hopper) will fail if they are
interrupted with a signal and this provides a workaround.

We would prefer to turn off sampling in all threads.  But that's more
complicated because it would have to be done from inside a signal
handler and that can't really be done safely.