Add a patch to libmonitor to add client support functions monitor_disable_new_threads() and monitor_enable_new_threads() to disable and reenable monitoring of side threads.
This allows hpcrun to work around a problem where CUDA creates a side
thread and then blocks until that thread finishes initializing, thus
leading to deadlock.

This patch is currently only useful for the GPU branches.  But it's a
noop if the client doesn't call thes support functions, so it's pretty
safe.