Add a new event REALTIME based on CLOCK_REALTIME and SIGEV_THREAD_ID to itimer.c.  This has two major differences with WALLCLOCK (itimer). CLOCK_REALTIME runs while blocked in the kernel, itimer does not. Also, the SIGEV_THREAD_ID extension guarantees per-thread signal delivery, while itimer is process-wide.
SIGEV_THREAD_ID is a Linux extension, so test for this in configure.

Cleanup some of the spaghetti #ifdefs in itimer.c.  Drop support for
Catamount (it doesn't compile anyway).  It looks like a long diff, but
it's mostly just moving code around and factoring things to support
two events.

Note: we don't allow running WALLCLOCK and REALTIME together.