Add proxy sampling for derived PAPI events.  That is, as long as some PAPI event supports hardware overflow, we can approximate derived events by reading the counters in the handler and taking the difference between two samples.  This is useful for new Intel CPUs where PAPI_FP_OPS is a derived event.
In my tests, the count of events is accurate, but the location of the
events is not so accurate.  For example, in a loop of mixed integer
and floating point where we sample on TOT_CYC and proxy sample on
FP_OPS, the FP_OPS might appear anywhere in the body of the loop.  But
as long as the loop is sufficiently long, the aggreate count for the
entire loop (and higher in the tree) will be correct.

No special options are needed for proxy sampling.  Just include the
derived events on the hpcrun command line as usual.  The threshold is
ignored.  A secret, special threshold of zero will force the event to
use proxy sampling (mostly for testing).