Fix hpclink to handle the cross-compile link problem of needing hpcfnbounds from the front-end tree and all the libraries, PAPI and CC from the back-end tree.
The solution here is to run hpclink from the back-end install with an
option (--front-end) to specify the install directory (prefix) of the
front-end tree.  For example:

  hpclink --front-end /front/end/prefix cross-gcc file ...

The other way of running hpclink from the front-end tree with an
option for the back-end tree is possible but more complicated.  The
problem is that the back-end tree doesn't have a convenient way of
specifying CC or locating PAPI (not copied to ext-libs) without
writing these values to a config file.