On Blue Gene/Q, build libmonitor with the back-end powerpc64-bgq-linux-gcc compiler.  There is something in the code that the front-end gcc produces that breaks libmonitor's wrapping of main (hangs inside darshan, PAMI and MPI_Finalize), even with four arguments to main.
Add the configure option --enable-bgq for this.  Configure will
auto-detect Blue Gene/Q and use powerpc64-bgq-linux-gcc by default
(with a warning).  But use --enable-bgq for the back-end compute nodes
and --disable-bgq if building for the front end.

This may be a temporary solution and we may need to use the full cross
compile like we do for MIC.  But the powerpc64-bgq-linux compilers
currently break other externals's packages and toolkit over some
libtool 'static linking of dynamic object' issues.

With this and the previous libmonitor patch, I think we now run
correctly on Blue Gene/Q.  At least, I don't know an example where
it fails.