Adjust the method for adding a sequence of alien nodes for inlined code in determineContext().
This version adds one intermediate alien node for each step in the
inlined sequence from symtabAPI plus one final node where the code
comes from.  Also, we fake the procedure name to the source line
number to get the alien cache to handle the least common ancestor
correctly.

There is still more work to do, but this version seems to have the
right voodoo for producing the sequence of inlined steps.