Add thread local data (via gcc __thread construct) to ensure that everything is initialized.
This fixes an apparent bug with WALLCLOCK: A WALLCLOCK sample originating in 1
thread is intermittently being delivered to a different thread. If the thread to
which the sample is delivered has not yet been initialized, then segfault occurs.

At the moment, this is only a hypothesis.

The fix, however, does make the problem go away.

If a more principled solution is found, then it will replace this fix.