Add a work around for some broken paths related to libstdc++.la that affect xerces on some Xeon Phi systems.
On some newer Xeon Phi systems (phi.cs), the libdir path in the back
end (k1om) libstdc++.la is broken.  (It contains the back end path and
won't link from the front end.)  This propagates to dependency_libs in
xerces and breaks the build for hpcprof.

This patch is an approximate work around that removes libstdc++.la
from dependency_libs in libxerces-c.la in a cross compile.  Going
forward, we may need to adjust the fix if we see more examples of the
same problem.