Rework the libelf build scripts to use elfutils instead of the much older libelf-0.8.13.  Dyninst no longer runs cleanly with the old libelf.
Change the zlib default to build and remove the static only option.
Zlib is now a prerequisite for elfutils.

Rework the zlib config tests to allow a multilib lib directory.  This
allows using zlib from /usr on 64-bit systems.

At this point, libdwarf and dyninst should build cleanly with the new
elfutils in most cases.  However, the elfutils-0.167 tar file is not
yet committed (waiting on the make-lite script).

Also, the toolkit configure and make files don't yet understand
elfutils and the dependance on zlib.